Move telemetry_optin_notice_seen into the test suite that it depends on
